- refactor(app): batch 3 route-local _components/_lib for billing, contribute, leaderboard, draft
