+++ src/app/draft-history/leaderboard/_components/LeaderboardClient.tsx
@@ -352,4 +352,3 @@ function VerifiedCheck() {
     </TooltipProvider>
   );
 }
-
