+++ src/components/support/supportPromptTierPlans.ts
@@ -1,4 +1,4 @@
-import { SUPPORTER_TIER_PLANS } from "@/app/contribute/supporterTierPlans";
+import { SUPPORTER_TIER_PLANS } from "@/app/contribute/_lib/supporterTierPlans";
 
 export const SUPPORT_PROMPT_TIER_PLANS = SUPPORTER_TIER_PLANS;
